create add tag

+++ internal/models/event.go
@@ -10,5 +10,5 @@ type Event struct {
 	Description string    `json:"description" binding:"required"`
 	Location    string    `json:"location" binding:"required"`
 	DateTime    time.Time `json:"createTime"`
-	UserId      int       `json:"userId"`
+	UserId      int64     `json:"userId"`
 }
