+++ internal/models/file.go
@@ -8,12 +8,12 @@ type File struct {
 	Description string    `json:"description"`
 	URL         string    `json:"url" binding:"required"`
 	DateTime    time.Time `json:"createTime"`
-	UserId      int       `json:"userId"`
+	UserId      int64     `json:"userId"`
 }
 
 type Tag struct {
 	ID       int64     `json:"id"`
 	Name     string    `json:"name" binding:"required"`
 	DateTime time.Time `json:"createTime"`
-	UserId   int       `json:"userId"`
+	UserId   int64     `json:"userId"`
 }
